changed ID to long type

+++ koty-application/src/main/java/pl/kobietydokodu/koty/dao/CatDAO.java
@@ -6,7 +6,7 @@ import pl.kobietydokodu.koty.domain.Cat;
 public interface CatDAO {
 
 	public List<Cat> findAll();
-	public Cat findById(Integer id);
+	public Cat findById(Long id);
 	public void add(Cat kot);
 	public void edit(Cat kot);
 	public void delete(Long idKot);
