+++ koty-service/src/main/java/pl/kobietydokodu/koty/service/CatService.java
@@ -6,7 +6,7 @@ import pl.kobietydokodu.koty.domain.Cat;
 public interface CatService {
 	
 	public List<Cat> findAll();
-	public Cat findById(Integer id);
+	public Cat findById(Long id);
 	public void add(Cat kot);
 	public void edit(Cat kot);
 	public void delete(Long idKot);
